--- e2e-tests/bonding_test.go
@@ -1,30 +1,0 @@
-package e2e_test
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-
-	consumertypes "github.com/cosmos/interchain-security/x/ccv/consumer/types"
-	providertypes "github.com/cosmos/interchain-security/x/ccv/provider/types"
-)
-
-// TestPacketRoundtrip tests a CCV packet roundtrip when tokens are bonded on provider
-func (s *ProviderTestSuite) TestPacketRoundtrip() {
-	s.SetupCCVChannel()
-
-	// Bond some tokens on provider to change validator powers
-	bondAmt := sdk.NewInt(1000000)
-	delAddr := s.providerChain.SenderAccount.GetAddress()
-	delegate(s, delAddr, bondAmt)
-
-	// Send CCV packet to consumer
-	s.providerChain.NextBlock()
-
-	// Relay 1 VSC packet from provider to consumer
-	relayAllCommittedPackets(s, s.providerChain, s.path, providertypes.PortID, s.path.EndpointB.ChannelID, 1)
-
-	// Increment time so that the unbonding period ends on the provider
-	incrementTimeByUnbondingPeriod(s, Provider)
-
-	// Relay 1 VSCMatured packet from consumer to provider
-	relayAllCommittedPackets(s, s.consumerChain, s.path, consumertypes.PortID, s.path.EndpointA.ChannelID, 1)
-}
