+++ e2e-tests/distribution_test.go
@@ -1,0 +1,3 @@
+package e2e_test
+
+// TODO: Add tests for the reward distribution sub protocol
